refactor(combat): extract shared turn-progress helpers to src/utils/
Deduplicate getTurnOrderIndexForCombatant and hasCombatantTurnEndedThisRound
from minionGroupTokenBadges.ts and combat.utils.ts into a single shared
utility, fixing an id-access inconsistency between the two copies.
